Add `BindingSourceKeyExtensions` to centralize `BindingSource` lookup
